Modularize focus functionality
